fix: widen anomaly mttd columns

+++ src/storage/schema.sql
@@ -66,8 +66,9 @@ CREATE TABLE ANOMALIES (
     RULE_IDS             NCLOB,
     TOTAL_REQUESTS       INTEGER,
     ERROR_RATE           DECIMAL(5, 4),
-    PIPELINE_MTTD_MS     INTEGER,                    -- detected_at - ingested_at
-    E2E_MTTD_MS          INTEGER,                    -- detected_at - log event time
+    -- BIGINT keeps long-running/backfilled events from overflowing HANA INTEGER.
+    PIPELINE_MTTD_MS     BIGINT,                     -- detected_at - ingested_at
+    E2E_MTTD_MS          BIGINT,                     -- detected_at - log event time
     ALERT_ID             NVARCHAR(100),              -- idempotency key
     DEDUP_KEY            NVARCHAR(100),
     WEBHOOK_SENT         BOOLEAN         DEFAULT FALSE,
